+++ runtime/shared/__mocks__/fbt.js
@@ -96,11 +96,20 @@ function unwrap(json) {
   return payload;
 }
 
-fbt._ = jest.fn().mockImplementation((wrappedJSON, args) => {
-  const unwrappedJson = unwrap(wrappedJSON);
-  const jsfbt = unwrappedJson.jsfbt;
-  return fbtRuntime._(unwrappedJson.type === 'text' ? jsfbt : jsfbt.t, args);
-});
+fbt._ = jest.fn().mockImplementation(
+  // Do not use arrow function because we need `this` to refer to whatever runtime context
+  // we'll be having later.
+  // We expect `this` to be either from the usual fbt.js or from fbs.js
+  function (wrappedJSON, args) {
+    const unwrappedJson = unwrap(wrappedJSON);
+    const jsfbt = unwrappedJson.jsfbt;
+    return fbtRuntime._.call(
+      this,
+      unwrappedJson.type === 'text' ? jsfbt : jsfbt.t,
+      args,
+    );
+  },
+);
 
 fbt._.getCallString = index => unwrap(fbt._.mock.calls[index][0]);
 
@@ -112,6 +121,7 @@ fbt._.getCallString = index => unwrap(fbt._.mock.calls[index][0]);
   '_plural',
   '_pronoun',
   '_subject',
+  '_wrapContent',
   'isFbtInstance',
 ].forEach(methodName => {
   invariant(
@@ -119,7 +129,7 @@ fbt._.getCallString = index => unwrap(fbt._.mock.calls[index][0]);
     'Expected method fbt.%s() to be defined',
     methodName,
   );
-  fbt[methodName] = (...args) => fbtRuntime[methodName](...args);
+  fbt[methodName] = fbtRuntime[methodName];
 });
 
 module.exports = fbt;