+++ CHANGELOG.md
@@ -53,6 +53,8 @@ List of changes for each released npm package version.
     Unreleased changes that have landed in master. Click to see more.
   </summary>
 
+  - [fix] Force fbs:plural to only show plain text custom UI values.
+
 </details>
 
 - 0.16.6: